requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.8.6
pymongo==4.6.0
motor==3.3.1
python-telegram-bot==20.8
selenium==4.15.2
webdriver-manager==4.0.1
apscheduler==3.10.4
email-validator==2.0.0
httpx==0.24.1
bs4==0.0.1
sqlalchemy==2.0.27
PyPDF2==3.0.1
python-docx==1.0.1
google-ads==22.1.0
psycopg2-binary==2.9.9
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.105.0 